servertest: avoid duplicates
Commit 351a0ce3d2d972e8377c83013f1759ac3d08f1cd, which switched from
QList to QVector, also switched a QSet to QVector: this gives as result
duplicate items for the connection results.

Since having duplicate items does not make sense, then switch the
internal storage back to QSet, manually converting it to the QVector
that is emitted as parameter of the finished signal.