Remove MailTransport dependency from MailTransportAkonadi
This removes two things:
- The transport() helper method in TransportAttribute. This is only used
in the mail dispatcher agent and easy to replace.
- The transport validity check when queuing an email. This check is
repeated by the dispatcher agent at the time of sending again anyway
(and that's the more relevant time, as transport configuration can change
meanwhile).

As a result both libraries are now independent of each other. This might
seem a bit weird, but to me this shows that what is in MailTransportAkonadi
probably rather belongs into AkonadiMime instead. That's for a subsequent
change though.