Change default encryption to SSL/TLS
This also adapt the default port used for SMTP as 25 is the standard
port for non-encrypted communication and 587 is the default port for
encrypted SMTP communication.

Also improve the description of the key as the enum value SSL and TLS
are incorectly named. TLS is actually STARTTLS and SSL is TLS. We should
rename the enums later.

See https://www.cloudflare.com/learning/email-security/smtp-port-25-587/
and https://serverfault.com/a/648282